+++ dto/accountResponse.go
@@ -1,9 +1,9 @@
 package dto
 
 type AccountResponse struct {
-	Id           int64  `json:"id"`
-	CustomerId   int64  `json:"customerId"`
-	PortfolioId  int64  `json:"portfolioId"`
+	Id           int    `json:"id"`
+	CustomerId   int    `json:"customerId"`
+	PortfolioId  int    `json:"portfolioId"`
 	CurrencyCode string `json:"currencyCode"`
-	Balance      int64  `json:"balance"`
+	Balance      int    `json:"balance"`
 }
