+++ dto/assetResponse.go
@@ -1,8 +1,8 @@
 package dto
 
 type AssetResponse struct {
-	Id      int64  `json:"id"`
+	Id      int    `json:"id"`
 	Isin    string `json:"isin"`
 	Name    string `json:"name"`
-	Percent int64  `json:"percent"`
+	Percent int    `json:"percent"`
 }
