+++ dto/instructionResponse.go
@@ -1,11 +1,11 @@
 package dto
 
 type InstructionResponse struct {
-	Id              int64  `json:"id"`
-	OrderSheetId    int64  `json:"orderSheetId"`
+	Id              int    `json:"id"`
+	OrderSheetId    int    `json:"orderSheetId"`
 	InstructionType string `json:"instructionType"`
 	Isin            string `json:"isin"`
-	Amount          int64  `json:"amount"`
+	Amount          int    `json:"amount"`
 	CurrencyCode    string `json:"currencyCode"`
 	CreatedAt       string `json:"createdAt"`
 }
