+++ dto/orderSheetResponse.go
@@ -1,8 +1,8 @@
 package dto
 
 type OrderSheetResponse struct {
-	Id                  int64                  `json:"id"`
-	WithdrawalRequestId int64                  `json:"withdrawalRequestId"`
+	Id                  int                    `json:"id"`
+	WithdrawalRequestId int                    `json:"withdrawalRequestId"`
 	Status              string                 `json:"status"`
 	CreatedAt           string                 `json:"createdAt"`
 	Instructions        *[]InstructionResponse `json:"instructions"`
